Extracting the text from the code
