+++ FrontEnd/src/pages/ProfilePage/ProfilePage.module.css
@@ -3,5 +3,4 @@
     flex-direction: column;
     padding-top: 32px;
     left: 192px;
-    flex-grow: 1;
 }
