+++ FrontEnd/src/pages/ProfilePage/ProfilePageComponents/TextFormNames.js
@@ -1,0 +1,12 @@
+const FORM_NAMES = [
+    'UserInfo',
+    'GeneralInfo',
+    'ContactsInfo',
+    'ProductServiceInfo',
+    'AdditionalInfo',
+    'StartupInfo',
+    'Delete',
+    'ChangePassword'
+];
+
+export default FORM_NAMES;
