+++ FrontEnd/src/pages/ProfilePage/ProfilePageComponents/TextInfoLinks.js
@@ -1,0 +1,31 @@
+const INFOLINKS = [
+    {
+        title: 'Інформація про користувача',
+        link: '/user-info',
+    },
+    {
+        title: 'Загальна інформація',
+        link: '/general-info',
+    },
+    {
+        title: 'Контакти',
+        link: '/contacts',
+    },
+    {
+        title: 'Інформація про товари/ послуги',
+        link: '/products-service-info',
+        tooltipText: 'Цей розділ доступний, коли обрано опцію "Зареєстрована компанія" у розділі Загальна інформація'
+    },
+    {
+        title: 'Додаткова інформація',
+        link: '/additional-info',
+        tooltipText: 'Цей розділ доступний, коли обрано опцію "Зареєстрована компанія" у розділі Загальна інформація'
+    },
+    {
+        title: 'Стартап',
+        link: '/startup',
+        tooltipText: 'Цей розділ доступний, коли обрано опцію "Стартап проект" у розділі Загальна інформація'
+    },
+];
+
+export default INFOLINKS;
